Part 2
So the form is fully reformed with new visual settings as well as added confirmation message box which will help in avoiding accidental deletion of the customer accounts

+++ BankGG-windowsApp/BankGG/DeleteAccount.cs
@@ -26,9 +26,17 @@ namespace BankGG
             InitializeComponent();
         }
 
-        private async void SearchBtn_Click(object sender, EventArgs e)
-        {
+     
 
+        private void Form5_Load(object sender, EventArgs e)
+        {
+            customer = new FireSharp.FirebaseClient(config);
+        }
+
+     
+
+        private async void SearchBtn_Click_1(object sender, EventArgs e)
+        {
             try
             {
                 FirebaseResponse responseGetdata = await customer.GetTaskAsync("Bankdetails/" + maskedTextBox1.Text); ;
@@ -37,60 +45,60 @@ namespace BankGG
 
 
 
-                DataTable accountinfo = new DataTable();
-                accountinfo.Columns.Add("Account no");
-                accountinfo.Columns.Add("Name");
-                accountinfo.Columns.Add("Birth Date");
-                accountinfo.Columns.Add("Email");
-                accountinfo.Columns.Add("Address");
-                accountinfo.Columns.Add("Amount");
-                accountinfo.Columns.Add("Account Type");
 
-                DataRow InfoRow = accountinfo.NewRow();
-                InfoRow["Account no"] = accountdata.ID;
-                InfoRow["Name"] = accountdata.Name;
-                InfoRow["Birth Date"] = accountdata.BirthDate;
-                InfoRow["Email"] = accountdata.Email;
-                InfoRow["Address"] = accountdata.address;
-                InfoRow["Amount"] = accountdata.Amount;
-                InfoRow["Account Type"] = accountdata.AccountType;
-                accountinfo.Rows.Add(InfoRow);
+
+
+
                 amount = Convert.ToInt32(accountdata.Amount);
-                dataGridView1.Visible = true;
-                dataGridView1.DataSource = accountinfo;
+                Acconolbl.Visible = false;
+                maskedTextBox1.Visible = false;
                 SearchBtn.Visible = false;
-                dataGridView1.Visible = true;
+                AccNamelbl.Visible = true;
+                AccountNameislbl.Visible = true;
                 DeleteBtn.Visible = true;
+                AccoNo.Visible = true;
+                AccountNoislbl.Visible = true;
+                AccoNo.Text = accountdata.ID.ToString();
+                AccNamelbl.Text = accountdata.Name;
             }
-            catch(NullReferenceException es)
+            catch (NullReferenceException es)
             {
                 MessageBox.Show("Account not found or an network is unstable");
             }
-            
         }
 
-        private void Form5_Load(object sender, EventArgs e)
+        private async void DeleteBtn_Click_1(object sender, EventArgs e)
         {
-            customer = new FireSharp.FirebaseClient(config);
-        }
-
-        private async void  DeleteBtn_Click(object sender, EventArgs e)
-        {
-            if (amount==0)
+            if(amount == 0)
             {
                 try
                 {
-                    FirebaseResponse response1 = await customer.DeleteTaskAsync("Bankdetails/" + maskedTextBox1.Text);
-                    MessageBox.Show("Account Deleted");
-                }catch(NullReferenceException eg)
+                    if(MessageBox.Show("Are you sure You want to delete your account", "Confirmation", MessageBoxButtons.OKCancel)==DialogResult.OK)
+                    {
+                        FirebaseResponse response1 = await customer.DeleteTaskAsync("Bankdetails/" + maskedTextBox1.Text);
+                        MessageBox.Show("Account Deleted");
+                    }
+                    else
+                    {
+                        MessageBox.Show("Account deletion operation was Cancelled by the user");
+                    }
+                    
+                   
+                }
+                catch (NullReferenceException eg)
                 {
                     MessageBox.Show("Could not delete the account Operation failed");
                 }
             }
-            else 
+            else
             {
                 MessageBox.Show("Account balance is not Zero for account to be delete please withdraw all cash for Account to be deleted");
             }
+        }
+
+        private void guna2Button1_Click(object sender, EventArgs e)
+        {
+            this.Close();
         }
     }
 }
